Clean up vets info page
