+++ src/applications/disability-benefits/996/pages/veteranInformation.js
@@ -1,63 +1,16 @@
-import React from 'react';
-import { connect } from 'react-redux';
-import moment from 'moment';
+import veteranInfoDescription from '../content/veteranInformation';
 
-import { srSubstitute } from '../../all-claims/utils';
-import { genderLabels } from 'platform/static-data/labels';
-
-const mask = srSubstitute('●●●–●●–', 'ending with');
-
-export const veteranInfoView = ({ profile = {}, veteran = {} }) => {
-  const { ssnLastFour, vaFileNumber } = veteran;
-  const { dob, gender, userFullName } = profile;
-  const { first = '', middle = '', last = '', suffix = '' } = userFullName;
-  return (
-    <>
-      <p>This is the personal information we have on file for you.</p>
-      <br />
-      <div className="blue-bar-block">
-        <strong className="name">
-          {first} {middle} {last} {suffix}
-        </strong>
-        {ssnLastFour && (
-          <p className="ssn">
-            Social Security number: {mask} {ssnLastFour.slice(-4)}
-          </p>
-        )}
-        {vaFileNumber && (
-          <p className="vafn">
-            VA file number: {mask} {vaFileNumber.slice(-4)}
-          </p>
-        )}
-        <p>
-          Date of birth:{' '}
-          <span className="dob">{dob ? moment(dob).format('LL') : ''}</span>
-        </p>
-        <p>
-          Gender:{' '}
-          <span className="gender">
-            {(gender && genderLabels[gender]) || ''}
-          </span>
-        </p>
-      </div>
-      <br />
-      <p>
-        <strong>Note:</strong> If you need to update your personal information,
-        please call Veterans Benefits Assistance toll free at{' '}
-        <a className="nowrap" href="tel:1-800-827-1000">
-          800-827-1000
-        </a>
-        , Monday through Friday, 8:00 a.m. to 9:00 p.m. ET.
-      </p>
-    </>
-  );
+const veteranInformation = {
+  uiSchema: {
+    'ui:description': veteranInfoDescription,
+    'ui:options': {
+      hideOnReview: true,
+    },
+  },
+  schema: {
+    type: 'object',
+    properties: {},
+  },
 };
 
-export const veteranInfoDescription = connect(state => {
-  const profile = state.user?.profile;
-  const veteran = state.form?.loadedData?.formData;
-  return {
-    profile,
-    veteran,
-  };
-})(veteranInfoView);
+export default veteranInformation;
